// wls-gui-wahllokalsystem/src/api/wls-clients/generated-infomanagement-api/api.ts
export interface KonfigurationDTO {
    /**
     * 
     * @type {string}
     * @memberof KonfigurationDTO
     */
    'schluessel': string;
    /**
     * 
     * @type {string}
     * @memberof KonfigurationDTO
     */
    'wert'?: string;
    /**
     * 
     * @type {string}
     * @memberof KonfigurationDTO
     */
    'beschreibung'?: string;
    /**
     * 
     * @type {string}
     * @memberof KonfigurationDTO
     */
    'standardwert'?: string;
}